[Add]CalendarAndDrinkingRecordViewの大枠のレイアウト作成
